Refactor graph.rs into multiple files
