+++ src/graph.rs
@@ -31,10 +31,21 @@
 #[cfg(loom)]
 use loom::sync::atomic::Ordering;
 
+pub mod connection;
+mod graph_gen;
+mod node;
+pub mod run_graph;
+pub use connection::Connection;
+use connection::ConnectionError;
+use node::Node;
+pub use node::NodeBufferRef;
+pub use run_graph::{RunGraph, RunGraphSettings};
+
 use crate::scheduling::{MusicalTime, MusicalTimeMap};
 use crate::time::Superseconds;
+use rtrb::RingBuffer;
+use slotmap::{new_key_type, SecondaryMap, SlotMap};
 
-// #[cfg(not(loom))]
 use std::cell::UnsafeCell;
 use std::collections::{HashMap, HashSet};
 use std::mem;
@@ -214,12 +225,6 @@ impl GraphInput {
     }
 }
 
-// graph.connect(sine.to(GraphOut))
-// graph.connect(sine.to(GRAPH_OUT))
-// graph.connect(sine.to_graph_out())
-// graph.connect(GraphIn::to(sine).from_index(1).to_label("freq"))
-// graph.connect(GRAPH_IN.to(sine).from_index(1).to_label("freq"))
-
 #[derive(Clone)]
 pub struct ParameterChange {
     pub time: TimeKind,
@@ -283,361 +288,6 @@ pub enum TimeKind {
     Immediately,
 }
 
-/// Connection provides a convenient API for creating connections between nodes in a
-/// graph. When used for nodes in a running graph, the shadow engine will translate the
-/// Connection to a RtConnection which contains the full Path for finding the correct Graph
-/// as fast as possible.
-///
-/// A node can have any number of connections to/from other nodes or outputs.
-/// Multiple constant values will result in the sum of all the constants.
-#[derive(Clone)]
-pub enum Connection {
-    /// node to node
-    Node {
-        source: NodeAddress,
-        from_index: Option<usize>,
-        from_label: Option<&'static str>,
-        sink: NodeAddress,
-        /// if input_index and input_label are both None, the default is index 0
-        to_index: Option<usize>,
-        to_label: Option<&'static str>,
-        /// default: 1
-        channels: usize,
-        feedback: bool,
-    },
-    /// constant to node
-    Constant {
-        value: Sample,
-        sink: Option<NodeAddress>,
-        to_index: Option<usize>,
-        to_label: Option<&'static str>,
-    },
-    /// node to graph output
-    GraphOutput {
-        source: NodeAddress,
-        from_index: Option<usize>,
-        from_label: Option<&'static str>,
-        to_index: usize,
-        channels: usize,
-    },
-    /// graph input to node
-    GraphInput {
-        sink: NodeAddress,
-        from_index: usize,
-        /// if input_index and input_label are both None, the default is index 0
-        to_index: Option<usize>,
-        to_label: Option<&'static str>,
-        channels: usize,
-    },
-    Clear {
-        node: NodeAddress,
-        /// connections to this node
-        input_nodes: bool,
-        /// constant input values of this node
-        input_constants: bool,
-        /// connections from this node to other nodes
-        output_nodes: bool,
-        /// connections from this node to the graph output(s)
-        graph_outputs: bool,
-        /// connections from the graph inputs to the node
-        graph_inputs: bool,
-    },
-}
-
-pub fn constant(value: Sample) -> Connection {
-    Connection::Constant {
-        value,
-        sink: None,
-        to_index: None,
-        to_label: None,
-    }
-}
-impl Connection {
-    pub fn graph_output(source_node: &NodeAddress) -> Self {
-        Self::GraphOutput {
-            source: source_node.clone(),
-            from_index: Some(0),
-            from_label: None,
-            to_index: 0,
-            channels: 1,
-        }
-    }
-    pub fn graph_input(sink_node: &NodeAddress) -> Self {
-        Self::GraphInput {
-            sink: sink_node.clone(),
-            from_index: 0,
-            to_index: None,
-            to_label: None,
-            channels: 1,
-        }
-    }
-    pub fn clear_constants(node: &NodeAddress) -> Self {
-        Self::Clear {
-            node: node.clone(),
-            input_constants: true,
-            input_nodes: false,
-            output_nodes: false,
-            graph_outputs: false,
-            graph_inputs: false,
-        }
-    }
-    pub fn clear_from_nodes(node: &NodeAddress) -> Self {
-        Self::Clear {
-            node: node.clone(),
-            input_constants: false,
-            input_nodes: true,
-            output_nodes: false,
-            graph_outputs: false,
-            graph_inputs: false,
-        }
-    }
-    pub fn clear_to_nodes(node: &NodeAddress) -> Self {
-        Self::Clear {
-            node: node.clone(),
-            input_constants: false,
-            input_nodes: false,
-            output_nodes: true,
-            graph_outputs: false,
-            graph_inputs: false,
-        }
-    }
-    pub fn clear_to_graph_outputs(node: &NodeAddress) -> Self {
-        Self::Clear {
-            node: node.clone(),
-            input_constants: false,
-            input_nodes: false,
-            output_nodes: false,
-            graph_outputs: true,
-            graph_inputs: false,
-        }
-    }
-    pub fn clear_from_graph_inputs(node: &NodeAddress) -> Self {
-        Self::Clear {
-            node: node.clone(),
-            input_constants: false,
-            input_nodes: false,
-            output_nodes: false,
-            graph_outputs: false,
-            graph_inputs: true,
-        }
-    }
-    /// Sets the source of a Connection. Only valid for Connection::Node and
-    /// Connection::GraphOutput. On other variants it does nothing.
-    pub fn from(mut self, source_node: &NodeAddress) -> Self {
-        match &mut self {
-            Connection::Node { ref mut source, .. } => {
-                *source = source_node.clone();
-            }
-            Connection::GraphInput { .. } => {}
-            Connection::Constant { .. } => {}
-            Connection::GraphOutput { ref mut source, .. } => {
-                *source = source_node.clone();
-            }
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    /// Sets the source of a Connection. Does nothing on a
-    /// Connection::GraphOutput or Connection::Clear.
-    pub fn to(mut self, sink_node: &NodeAddress) -> Self {
-        match &mut self {
-            Connection::Node { ref mut sink, .. } => {
-                *sink = sink_node.clone();
-            }
-            Connection::GraphInput { ref mut sink, .. } => {
-                *sink = sink_node.clone();
-            }
-            Connection::Constant { ref mut sink, .. } => {
-                *sink = Some(sink_node.clone());
-            }
-            Connection::GraphOutput { .. } => {}
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    pub fn to_label(mut self, label: &'static str) -> Self {
-        match &mut self {
-            Connection::Node {
-                to_label: input_label,
-                to_index,
-                ..
-            } => {
-                *input_label = Some(label);
-                *to_index = None;
-            }
-            Connection::Constant {
-                to_label: input_label,
-                ..
-            } => {
-                *input_label = Some(label);
-            }
-            Connection::GraphOutput { .. } => {}
-            Connection::GraphInput {
-                to_label: input_label,
-                to_index,
-                ..
-            } => {
-                *input_label = Some(label);
-                *to_index = None;
-            }
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    /// Shortcut for `to_label`
-    pub fn tl(self, label: &'static str) -> Self {
-        self.to_label(label)
-    }
-    pub fn to_index(mut self, index: usize) -> Self {
-        match &mut self {
-            Connection::Node {
-                to_index: input_index,
-                to_label,
-                ..
-            } => {
-                *input_index = Some(index);
-                *to_label = None;
-            }
-            Connection::Constant {
-                to_index: input_index,
-                ..
-            } => {
-                *input_index = Some(index);
-            }
-            Connection::GraphOutput { to_index, .. } => {
-                *to_index = index;
-            }
-            Connection::Clear { .. } => {}
-            Connection::GraphInput {
-                to_index, to_label, ..
-            } => {
-                *to_index = Some(index);
-                *to_label = None;
-            }
-        }
-        self
-    }
-    pub fn get_from_index(&self) -> Option<usize> {
-        match &self {
-            Connection::Node { from_index, .. } => *from_index,
-            Connection::GraphOutput { from_index, .. } => *from_index,
-            Connection::Clear { .. } => None,
-            Connection::Constant { .. } => None,
-            Connection::GraphInput { from_index, .. } => Some(*from_index),
-        }
-    }
-    pub fn get_to_index(&self) -> Option<usize> {
-        match &self {
-            Connection::Node {
-                to_index: input_index,
-                ..
-            } => *input_index,
-            Connection::Constant {
-                to_index: input_index,
-                ..
-            } => *input_index,
-            Connection::GraphOutput { to_index, .. } => Some(*to_index),
-            Connection::Clear { .. } => None,
-            Connection::GraphInput { to_index, .. } => *to_index,
-        }
-    }
-    /// Shortcut for `to_index`
-    pub fn ti(self, index: usize) -> Self {
-        self.to_index(index)
-    }
-    pub fn from_index(mut self, index: usize) -> Self {
-        match &mut self {
-            Connection::Node {
-                from_index,
-                from_label,
-                ..
-            } => {
-                *from_index = Some(index);
-                *from_label = None;
-            }
-            Connection::Constant { .. } => {}
-            Connection::GraphOutput {
-                from_index,
-                from_label,
-                ..
-            } => {
-                *from_index = Some(index);
-                *from_label = None;
-            }
-            Connection::GraphInput { from_index, .. } => {
-                *from_index = index;
-            }
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    pub fn from_label(mut self, label: &'static str) -> Self {
-        match &mut self {
-            Connection::Node {
-                from_index,
-                from_label,
-                ..
-            } => {
-                *from_label = Some(label);
-                *from_index = None;
-            }
-            Connection::Constant { .. } => {}
-            Connection::GraphOutput {
-                from_index,
-                from_label,
-                ..
-            } => {
-                *from_label = Some(label);
-                *from_index = None;
-            }
-            Connection::GraphInput { .. } => {}
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    /// Set how many channels should be connected together. This is most useful
-    /// for multi channel connections e.g. stereo -> stereo. Inputs/outputs will
-    /// wrap if the `num_channels` value if higher than the number of
-    /// inputs/outputs available.
-    pub fn channels(mut self, num_channels: usize) -> Self {
-        match &mut self {
-            Connection::Node { channels, .. } => {
-                *channels = num_channels;
-            }
-            Connection::Constant { .. } => {}
-            Connection::GraphOutput { channels, .. } => {
-                *channels = num_channels;
-            }
-            Connection::GraphInput { channels, .. } => {
-                *channels = num_channels;
-            }
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    pub fn feedback(mut self, activate: bool) -> Self {
-        match &mut self {
-            Connection::Node { feedback, .. } => {
-                *feedback = activate;
-            }
-            Connection::Constant { .. } => {}
-            Connection::GraphOutput { .. } => {}
-            Connection::GraphInput { .. } => {}
-            Connection::Clear { .. } => {}
-        }
-        self
-    }
-    pub fn get_source_node(&self) -> Option<NodeAddress> {
-        match self {
-            Connection::Node { source, .. } => Some(source.clone()),
-            Connection::GraphOutput { source, .. } => Some(source.clone()),
-            Connection::GraphInput { .. }
-            | Connection::Constant { .. }
-            | Connection::Clear { .. } => None,
-        }
-    }
-}
-
 /// One task to complete, for the node graph Safety: Uses raw pointers to nodes
 /// and buffers. A node and its buffers may not be touched from the Graph while
 /// a Task containing pointers to it is running. This is guaranteed by an atomic
@@ -696,7 +346,7 @@ impl Task {
         }
         // Copy all graph inputs
         for (graph_input_index, node_input_index) in &self.graph_inputs_to_copy {
-            for i in 0..self.input_buffers.block_size {
+            for i in 0..self.input_buffers.block_size() {
                 self.input_buffers.write(
                     graph_inputs.read(*graph_input_index, i),
                     *node_input_index,
@@ -706,12 +356,11 @@ impl Task {
         }
         if self.start_node_at_sample <= sample_time_at_block_start {
             // Process node
-            let mut outputs = NodeBufferRef {
-                buf: self.output_buffers_first_ptr,
-                num_channels: self.num_outputs,
-                block_size: self.block_size,
-                block_start_offset: 0,
-            };
+            let mut outputs = NodeBufferRef::new(
+                self.output_buffers_first_ptr,
+                self.num_outputs,
+                self.block_size,
+            );
             let ctx = GenContext {
                 inputs: &self.input_buffers,
                 outputs: &mut outputs,
@@ -727,12 +376,11 @@ impl Task {
                 self.block_size - (self.start_node_at_sample - sample_time_at_block_start) as usize;
 
             // Process node
-            let mut outputs = NodeBufferRef {
-                buf: self.output_buffers_first_ptr,
-                num_channels: self.num_outputs,
-                block_size: self.block_size,
-                block_start_offset: 0,
-            };
+            let mut outputs = NodeBufferRef::new(
+                self.output_buffers_first_ptr,
+                self.num_outputs,
+                self.block_size,
+            );
             let partial_inputs =
                 unsafe { self.input_buffers.to_partial_block_size(new_block_size) };
             let mut partial_outputs = unsafe { outputs.to_partial_block_size(new_block_size) };
@@ -762,32 +410,6 @@ struct OutputTask {
 }
 unsafe impl Send for OutputTask {}
 
-#[derive(thiserror::Error, Debug, PartialEq)]
-pub enum ConnectionError {
-    #[error("The nodes that you are trying to connect are in different graphs. Nodes can only be connected within a graph.")]
-    DifferentGraphs,
-    #[error("The graph containing the NodeAdress provided was not found. The node itself may or may not exist.")]
-    GraphNotFound,
-    #[error("The NodeAddress does not exist. The Node may have been freed already.")]
-    NodeNotFound,
-    #[error("The given input label (`{0}`) is not available for the given node.")]
-    InvalidInputLabel(&'static str),
-    #[error("The given output label (`{0}`) is not available for the given node.")]
-    InvalidOutputLabel(&'static str),
-    #[error("You are trying to connect a node to itself. This can only be done using a feedback connection.")]
-    SameNode,
-    #[error("The sink node for the connection is not set and is required.")]
-    SinkNotSet,
-    #[error("You are trying to connect to channels that don't exist, either through direct indexing or a too high `channels` value for the input.")]
-    ChannelOutOfBounds,
-    #[error("You are referencing a node which has not yet been pushed to a graph and therefore cannot be found.")]
-    SourceNodeNotPushed,
-    #[error("You are referencing a node which has not yet been pushed to a graph and therefore cannot be found.")]
-    SinkNodeNotPushed,
-    #[error("The connection change required freeing a node, but the node could not be freed.")]
-    NodeFree(#[from] FreeError),
-}
-
 #[derive(thiserror::Error, Debug)]
 pub enum PushError {
     #[error("The target graph was not found. The GenOrGraph that was pushed is returned.")]
@@ -801,7 +423,7 @@ pub enum FreeError {
     #[error("The NodeAddress does not exist. The Node may have been freed already.")]
     NodeNotFound,
     #[error("The free action required making a new connection, but the connection failed.")]
-    ConnectionError(#[from] Box<ConnectionError>),
+    ConnectionError(#[from] Box<connection::ConnectionError>),
 }
 #[derive(thiserror::Error, Debug, PartialEq, Eq)]
 pub enum ScheduleError {
@@ -891,7 +513,7 @@ impl GenOrGraph for Graph {
         }
         // Create the GraphGen from the new Graph
         let gen = self.create_graph_gen().unwrap();
-        (Some(self), Box::new(gen))
+        (Some(self), gen)
     }
     fn into_gen_or_graph_enum(self) -> GenOrGraphEnum {
         GenOrGraphEnum::Graph(self)
@@ -930,208 +552,6 @@ pub trait Gen {
     }
 }
 
-/// Wrapper around a buffer of output samples. Panics at improper usage.
-///
-/// The number of channels corresponds to the number of inputs/outputs specified
-/// for the [`Gen`] this is given to. Trying to access channels outside of what
-/// has been declared will result in a panic at runtime.
-///
-/// The NodeBuffer does not own the data it points to, it is just a wrapper
-/// around a pointer to the buffer to make accessing it more convenient.
-///
-/// The samples are stored in a `*mut [Sample]` layout out one channel after the
-/// other as if it was a `*mut[*mut [Sample]]`. In benchmarks this was as fast as
-/// iterating over multiple `&mut[Sample]`
-///
-pub struct NodeBufferRef {
-    buf: *mut Sample,
-    num_channels: usize,
-    block_size: usize,
-    block_start_offset: usize,
-}
-impl NodeBufferRef {
-    /// Produces a buffer with 0 channels that you cannot read from or write to,
-    /// but it may be useful as a first input to a node that has no inputs.
-    pub fn null_buffer() -> Self {
-        Self {
-            buf: std::ptr::null_mut(),
-            num_channels: 0,
-            block_size: 0,
-            block_start_offset: 0,
-        }
-    }
-    /// Write a value to a single sample
-    #[inline]
-    pub fn write(&mut self, value: f32, channel: usize, sample_index: usize) {
-        assert!(channel < self.num_channels);
-        assert!(sample_index < self.block_size - self.block_start_offset);
-        assert!(!self.buf.is_null());
-        let buffer_slice = unsafe {
-            std::slice::from_raw_parts_mut(self.buf, self.num_channels * self.block_size)
-        };
-        unsafe {
-            *buffer_slice.get_unchecked_mut(
-                channel * self.block_size + sample_index + self.block_start_offset,
-            ) = value
-        };
-    }
-    /// Decouples the lifetime of the returned slice from self so that slices to
-    /// different channels can be returned.
-    ///
-    /// Safety: We guarantee that the channels don't overlap. Getting multiple
-    /// mutable references to the same channel this way is, however, UB. Use
-    /// `split_mut` for an iterator to all channels instead.
-    #[inline]
-    unsafe fn get_channel_mut<'a, 'b>(&'a mut self, channel: usize) -> &'b mut [f32] {
-        assert!(channel < self.num_channels);
-        assert!(!self.buf.is_null());
-        let channel_offset = channel * self.block_size;
-        let buffer_channel_slice = unsafe {
-            let ptr_at_channel = self.buf.add(channel_offset + self.block_start_offset);
-            std::slice::from_raw_parts_mut(ptr_at_channel, self.block_size)
-        };
-        buffer_channel_slice
-    }
-    #[inline]
-    pub fn get_channel(&self, channel: usize) -> &[f32] {
-        assert!(channel < self.num_channels);
-        assert!(!self.buf.is_null());
-        let channel_offset = channel * self.block_size;
-        let buffer_channel_slice = unsafe {
-            let ptr_at_channel = self.buf.add(channel_offset + self.block_start_offset);
-            std::slice::from_raw_parts(ptr_at_channel, self.block_size - self.block_start_offset)
-        };
-        buffer_channel_slice
-    }
-    /// Adds the given value to the selected sample
-    #[inline]
-    pub fn add(&mut self, value: f32, channel: usize, sample_index: usize) {
-        assert!(channel < self.num_channels);
-        assert!(sample_index < self.block_size - self.block_start_offset);
-        assert!(!self.buf.is_null());
-        unsafe {
-            let buffer_slice =
-                std::slice::from_raw_parts_mut(self.buf, self.num_channels * self.block_size);
-            let sample = (*buffer_slice).get_unchecked_mut(
-                channel * self.block_size + sample_index + self.block_start_offset,
-            );
-            *sample += value
-        };
-    }
-    /// Read one sample from a channel
-    #[inline]
-    pub fn read(&self, channel: usize, sample_index: usize) -> Sample {
-        assert!(channel < self.num_channels);
-        assert!(sample_index < self.block_size - self.block_start_offset);
-        assert!(!self.buf.is_null());
-        unsafe {
-            let buffer_slice =
-                std::slice::from_raw_parts_mut(self.buf, self.num_channels * self.block_size);
-            *(*buffer_slice)
-                .get_unchecked(channel * self.block_size + sample_index + self.block_start_offset)
-        }
-    }
-    /// Fill a channel with one and the same value
-    #[inline]
-    fn fill_channel(&mut self, value: f32, channel: usize) {
-        assert!(channel < self.num_channels);
-        assert!(!self.buf.is_null());
-        let channel_offset = channel * self.block_size;
-        let buffer_channel_slice = unsafe {
-            let ptr_at_channel = self.buf.add(channel_offset + self.block_start_offset);
-            std::slice::from_raw_parts_mut(
-                ptr_at_channel,
-                self.block_size - self.block_start_offset,
-            )
-        };
-        for sample in buffer_channel_slice {
-            *sample = value;
-        }
-    }
-    /// Fill all channels in the buffer with one and the same value
-    #[inline]
-    fn fill(&mut self, value: f32) {
-        assert!(!self.buf.is_null());
-        let buffer_slice = unsafe {
-            std::slice::from_raw_parts_mut(self.buf, self.num_channels * self.block_size)
-        };
-        for sample in buffer_slice {
-            *sample = value;
-        }
-    }
-    fn ptr_to_sample(&mut self, channel: usize, sample_index: usize) -> *mut Sample {
-        assert!(channel < self.num_channels);
-        assert!(sample_index < self.block_size);
-        unsafe {
-            self.buf
-                .add(channel * self.block_size + sample_index + self.block_start_offset)
-        }
-    }
-    /// returns the number of channels
-    pub fn channels(&self) -> usize {
-        self.num_channels
-    }
-    /// returns the block size for the buffer
-    pub fn block_size(&self) -> usize {
-        self.block_size - self.block_start_offset
-    }
-    /// Returns an iterator to the channels in this NodeBufferRef
-    pub fn split_mut(&mut self) -> NodeBufferRefSplitMut {
-        NodeBufferRefSplitMut {
-            buf: self.buf,
-            num_channels: self.num_channels,
-            block_size: self.block_size,
-            block_start_offset: self.block_start_offset,
-            _phantom_data: std::marker::PhantomData,
-        }
-    }
-    /// Create a version of self where the block is offset from the start. This
-    /// is useful for starting a node some way into a block.
-    ///
-    /// Safety: You mustn't use the original NodeBufferRef before this one is dropped.
-    unsafe fn to_partial_block_size(&mut self, new_block_size: usize) -> Self {
-        Self {
-            buf: self.buf,
-            num_channels: self.num_channels,
-            block_size: self.block_size,
-            block_start_offset: self.block_size - new_block_size,
-        }
-    }
-}
-/// A variant of a NodeBufferRef that can yield &mut[Sample] to all the channels.
-pub struct NodeBufferRefSplitMut<'a> {
-    buf: *mut Sample,
-    num_channels: usize,
-    block_size: usize,
-    block_start_offset: usize,
-    _phantom_data: std::marker::PhantomData<&'a mut ()>,
-}
-impl<'a> Iterator for NodeBufferRefSplitMut<'a> {
-    type Item = &'a mut [Sample];
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.num_channels == 0 {
-            return None;
-        }
-        assert!(!self.buf.is_null());
-        let buffer_channel_slice = unsafe {
-            let channel_start_ptr = self.buf.add(self.block_start_offset);
-            std::slice::from_raw_parts_mut(
-                channel_start_ptr,
-                self.block_size - self.block_start_offset,
-            )
-        };
-        unsafe {
-            self.buf = self.buf.add(self.block_size);
-        }
-        self.num_channels -= 1;
-        Some(buffer_channel_slice)
-    }
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.num_channels, Some(self.num_channels))
-    }
-}
-
 /// Gives access to the inputs and outputs buffers of a node for processing.
 pub struct GenContext<'a, 'b> {
     pub inputs: &'a NodeBufferRef,
@@ -1141,7 +561,7 @@ pub struct GenContext<'a, 'b> {
 impl<'a, 'b> GenContext<'a, 'b> {
     /// Returns the current block size
     pub fn block_size(&self) -> usize {
-        self.outputs.block_size
+        self.outputs.block_size()
     }
 }
 
@@ -1492,7 +912,7 @@ impl Graph {
     fn split_and_create_node(&mut self) -> Result<Node, String> {
         let block_size = self.block_size();
         let graph_gen = self.create_graph_gen()?;
-        let mut node = Node::new("graph", Box::new(graph_gen));
+        let mut node = Node::new("graph", graph_gen);
         node.init(block_size, self.sample_rate);
         self.recalculation_required = true;
         Ok(node)
@@ -2948,12 +2368,7 @@ impl Graph {
         let first_sample = self.inputs_buffers_ptr.ptr.cast::<f32>();
         for &node_key in &self.node_order {
             let num_inputs = nodes[node_key].num_inputs();
-            let mut input_buffers = NodeBufferRef {
-                buf: first_sample,
-                num_channels: num_inputs,
-                block_size: self.block_size,
-                block_start_offset: 0,
-            };
+            let mut input_buffers = NodeBufferRef::new(first_sample, num_inputs, self.block_size);
             // Collect inputs into the node's input buffer
             let input_edges = &self.node_input_edges[node_key];
             let graph_input_edges = &self.graph_input_edges[node_key];
@@ -2969,8 +2384,8 @@ impl Graph {
                     let from_channel = input_edge.from_output_index;
                     let to_channel = input_edge.to_input_index;
                     inputs_to_copy.push((
-                        output_values.ptr_to_sample(from_channel, sample_index),
-                        input_buffers.ptr_to_sample(to_channel, sample_index),
+                        unsafe { output_values.ptr_to_sample(from_channel, sample_index) },
+                        unsafe { input_buffers.ptr_to_sample(to_channel, sample_index) },
                     ));
                 }
             }
@@ -2983,24 +2398,18 @@ impl Graph {
                 let mut output_values = source.output_buffers();
                 for i in 0..self.block_size {
                     inputs_to_copy.push((
-                        output_values.ptr_to_sample(feedback_edge.from_output_index, i),
-                        input_buffers.ptr_to_sample(feedback_edge.from_output_index, i),
+                        unsafe { output_values.ptr_to_sample(feedback_edge.from_output_index, i) },
+                        unsafe { input_buffers.ptr_to_sample(feedback_edge.from_output_index, i) },
                     ));
                 }
             }
             let node = &nodes[node_key];
-            tasks.push(Task {
+            tasks.push(node.to_task(
                 node_key,
                 inputs_to_copy,
                 graph_inputs_to_copy,
                 input_buffers,
-                input_constants: node.input_constants,
-                gen: node.gen,
-                output_buffers_first_ptr: node.output_buffers_first_ptr,
-                block_size: node.block_size,
-                num_outputs: node.num_outputs,
-                start_node_at_sample: node.start_node_at_sample,
-            });
+            ));
         }
         tasks
     }
@@ -3019,7 +2428,7 @@ impl Graph {
     }
     /// Only one GraphGen can be created from a Graph, since otherwise nodes in
     /// the graph could be run multiple times.
-    fn create_graph_gen(&mut self) -> Result<GraphGen, String> {
+    fn create_graph_gen(&mut self) -> Result<Box<dyn Gen + Send>, String> {
         if self.graph_gen_communicator.is_some() {
             return Err(
                 "create_graph_gen: GraphGenCommunicator already existed for this graph".to_owned(),
@@ -3057,22 +2466,20 @@ impl Graph {
             timestamp: Arc::new(AtomicU64::new(0)),
         };
 
-        let graph_gen = GraphGen {
-            sample_rate: self.sample_rate,
-            current_task_data: task_data,
-            block_size: self.block_size,
-            num_outputs: self.num_outputs,
-            num_inputs: self.num_inputs,
-            graph_state: GenState::Continue,
-            sample_counter: 0,
-            timestamp: graph_gen_communicator.timestamp.clone(),
+        let graph_gen = graph_gen::make_graph_gen(
+            self.sample_rate,
+            task_data,
+            self.block_size,
+            self.num_outputs,
+            self.num_inputs,
+            graph_gen_communicator.timestamp.clone(),
             free_node_queue_producer,
             schedule_receiver,
-            _arc_nodes: self.nodes.clone(),
+            self.nodes.clone(),
             task_data_to_be_dropped_producer,
             new_task_data_consumer,
-            _arc_inputs_buffers_ptr: self.inputs_buffers_ptr.clone(),
-        };
+            self.inputs_buffers_ptr.clone(),
+        );
         self.graph_gen_communicator = Some(graph_gen_communicator);
         Ok(graph_gen)
     }
@@ -3180,215 +2587,6 @@ impl Graph {
 /// Node>>, but won't use it unless the Graph is dropped and it needs to keep
 /// the SlotMap alive, and the drop it when the GraphGen is dropped.
 unsafe impl Send for Graph {}
-
-impl Gen for GraphGen {
-    fn name(&self) -> &'static str {
-        "GraphGen"
-    }
-    fn process(&mut self, ctx: GenContext, resources: &mut Resources) -> GenState {
-        match self.graph_state {
-            GenState::Continue => {
-                // TODO: Support output with a different block size, i.e. local buffering and running this graph more or less often than the parent graph
-                //
-                //
-                let mut do_empty_buffer = None;
-                let mut do_mend_connections = None;
-                let num_new_task_data = self.new_task_data_consumer.slots();
-                if num_new_task_data > 0 {
-                    if let Ok(td_chunk) = self.new_task_data_consumer.read_chunk(num_new_task_data)
-                    {
-                        for td in td_chunk {
-                            // Setting `applied` to true signals that the new TaskData have been received and old data can be dropped
-                            td.applied.store(true, Ordering::SeqCst);
-                            let old_td = std::mem::replace(&mut self.current_task_data, td);
-                            match self.task_data_to_be_dropped_producer.push(old_td) {
-                            Ok(_) => (),
-                            Err(e) => eprintln!("RingBuffer for TaskData to be dropped was full. Please increase the size of the RingBuffer. The GraphGen will drop the TaskData here instead. e: {e}"),
-                        }
-                        }
-                    }
-                }
-
-                let task_data = &mut self.current_task_data;
-                let TaskData {
-                    applied: _,
-                    tasks,
-                    output_tasks,
-                } = task_data;
-
-                let changes = self.schedule_receiver.changes();
-
-                // Run the tasks
-                for task in tasks.iter_mut() {
-                    task.init_constants();
-                    // If there are any changes to the constants of the node, apply them here
-                    let mut i = 0;
-                    while i < changes.len() {
-                        let change = &changes[i];
-                        if change.key == task.node_key {
-                            let sample_to_apply = if change.timestamp < self.sample_counter {
-                                if change.timestamp != 0 {
-                                    // timestamps of 0 simply means as fast as possible. It is not an error or issue.
-                                    // TODO: Send this off of the audio thread
-                                    eprintln!("Warning: Scheduled change was applied late. Consider increasing latency.");
-                                }
-                                0
-                            } else {
-                                change.timestamp - self.sample_counter
-                            } as usize;
-                            if sample_to_apply < self.block_size {
-                                task.apply_constant_change(change, sample_to_apply);
-                                // TODO: This is inefficient since the the first
-                                // changes are the most likely to be removed,
-                                // and are the most expensive to remove. Either
-                                // the changes can be in reverse order, but then
-                                // pushing into the list always puts new changes
-                                // in the wrong place, or many changes can be
-                                // removed all at once after applying them.
-                                changes.remove(i);
-                            } else {
-                                i += 1;
-                            }
-                        } else {
-                            i += 1;
-                        }
-                    }
-                    match task.run(ctx.inputs, resources, self.sample_rate, self.sample_counter) {
-                        GenState::Continue => (),
-                        GenState::FreeSelf => {
-                            // We don't care if it fails since if it does the
-                            // node will still exist, return FreeSelf and get
-                            // added to the queue next block.
-                            self.free_node_queue_producer
-                                .push((task.node_key, GenState::FreeSelf))
-                                .ok();
-                        }
-                        GenState::FreeSelfMendConnections => {
-                            self.free_node_queue_producer
-                                .push((task.node_key, GenState::FreeSelfMendConnections))
-                                .ok();
-                        }
-                        GenState::FreeGraph(from_sample_nr) => {
-                            self.graph_state = GenState::FreeSelf;
-                            do_empty_buffer = Some(from_sample_nr);
-                        }
-                        GenState::FreeGraphMendConnections(from_sample_nr) => {
-                            self.graph_state = GenState::FreeSelfMendConnections;
-                            do_mend_connections = Some(from_sample_nr);
-                        }
-                    }
-                }
-
-                // Set the output of the graph
-                // Zero the output buffer.
-                ctx.outputs.fill(0.0);
-                for output_task in output_tasks.iter() {
-                    let input_values = output_task
-                        .input_buffers
-                        .get_channel(output_task.input_index);
-                    // Safety: We always drop the&mut refernce before requesting
-                    // another one so we cannot hold mutliple references to the
-                    // same channnel.
-                    let output =
-                        unsafe { ctx.outputs.get_channel_mut(output_task.graph_output_index) };
-                    for i in 0..self.block_size {
-                        let value = input_values[i];
-                        // Since many nodes may write to the same output we
-                        // need to add the outputs together. The Node makes no promise as to the content of
-                        // the output buffer provided.
-                        output[i] += value;
-                    }
-                }
-                if let Some(from_relative_sample_nr) = do_empty_buffer {
-                    for channel in ctx.outputs.split_mut() {
-                        for sample in &mut channel[from_relative_sample_nr..] {
-                            *sample = 0.0;
-                        }
-                    }
-                }
-                if let Some(from_relative_sample_nr) = do_mend_connections {
-                    let num_channels = ctx.inputs.channels().min(ctx.outputs.channels());
-                    for channel in 0..num_channels {
-                        let input = ctx.inputs.get_channel(channel);
-                        // Safety: We are only holding one &mut to a channel at a time.
-                        let output = unsafe { ctx.outputs.get_channel_mut(channel) };
-                        // TODO: Check if the input is constant or not first. Only non-constant inputs should be passed through (because it's "mending" the connection)
-                        for (i, o) in input[from_relative_sample_nr..]
-                            .iter()
-                            .zip(output[from_relative_sample_nr..].iter_mut())
-                        {
-                            *o = *i;
-                        }
-                    }
-                }
-                self.sample_counter += self.block_size as u64;
-                self.timestamp.store(self.sample_counter, Ordering::SeqCst);
-            }
-            GenState::FreeSelf => {
-                ctx.outputs.fill(0.0);
-            }
-            GenState::FreeSelfMendConnections => {
-                let num_channels = ctx.inputs.channels().min(ctx.outputs.channels());
-                for channel in 0..num_channels {
-                    let input = ctx.inputs.get_channel(channel);
-                    // Safety: We are only holding a &mut to one channel at a time.
-                    let output = unsafe { ctx.outputs.get_channel_mut(channel) };
-                    // TODO: Check if the input is constant or not first. Only non-constant inputs should be passed through (because it's "mending" the connection)
-                    for (i, o) in input.iter().zip(output.iter_mut()) {
-                        *o = *i;
-                    }
-                }
-            }
-            // These are unreachable because they are converted to the FreeSelf
-            // variants at the Graph level which is this level.
-            GenState::FreeGraph(_) => unreachable!(),
-            GenState::FreeGraphMendConnections(_) => unreachable!(),
-        }
-        self.graph_state
-    }
-    fn num_inputs(&self) -> usize {
-        self.num_inputs
-    }
-    fn num_outputs(&self) -> usize {
-        self.num_outputs
-    }
-}
-
-/// This gets placed as a dyn Gen in a Node in a Graph. It's how the Graph gets
-/// run. The Graph communicates with the GraphGen in a thread safe way.
-///
-/// # Safety
-/// Using this struct is safe only if used in conjunction with the
-/// Graph. The Graph owns nodes and gives its corresponding GraphGen raw
-/// pointers to them through Tasks, but it never accesses or deallocates a node
-/// while it can be accessed by the [`GraphGen`] through a Task. The [`GraphGen`]
-/// mustn't use the _arc_nodes field; it is only there to make sure the nodes
-/// don't get dropped.
-struct GraphGen {
-    block_size: usize,
-    sample_rate: Sample,
-    num_outputs: usize,
-    num_inputs: usize,
-    current_task_data: TaskData,
-    // This Arc is cloned from the Graph and exists so that if the Graph gets dropped, the GraphGen can continue on without segfaulting.
-    _arc_nodes: Arc<UnsafeCell<SlotMap<NodeKey, Node>>>,
-    // This Arc makes sure the input buffers allocation is valid for as long as it needs to be
-    _arc_inputs_buffers_ptr: Arc<OwnedRawBuffer>,
-    graph_state: GenState,
-    /// Stores the number of completed samples, updated at the end of a block
-    sample_counter: u64,
-    timestamp: Arc<AtomicU64>,
-    schedule_receiver: ScheduleReceiver,
-    free_node_queue_producer: rtrb::Producer<(NodeKey, GenState)>,
-    task_data_to_be_dropped_producer: rtrb::Producer<TaskData>,
-    new_task_data_consumer: rtrb::Consumer<TaskData>,
-}
-
-/// Safety: This impl of Send is required because of the Arc<UnsafeCell<...>> in
-/// GraphGen. The _arc_nodes field of GraphGen exists only so that the nodes
-/// won't get dropped if the Graph is dropped. The UnsafeCell will never be used
-/// to access the data from within GraphGen.
-unsafe impl Send for GraphGen {}
 
 /// The internal representation of a scheduled change to a running graph. This
 /// is what gets sent to the GraphGen.
@@ -3707,318 +2905,6 @@ impl GraphGenCommunicator {
         }
     }
 }
-
-/// Node is a very unsafe struct. Be very careful when changing it.
-///
-/// - The Gen is not allowed to be replaced.
-/// - The input_constants buffer mustn't be deallocated until the Node is dropped.
-/// - init must not be called after the Node has started being used on the audio thread.
-/// - the number of inputs/outputs of the Gen must not change.
-///
-/// It also must not be dropped while a Task exists with pointers to the buffers
-/// of the Node. Graph has a mechanism to ensure this.
-struct Node {
-    // name: &'static str,
-    /// index by input_channel
-    input_constants: *mut [Sample],
-    /// index by `output_channel * block_size + sample_index`
-    output_buffers: *mut [Sample],
-    output_buffers_first_ptr: *mut Sample,
-    block_size: usize,
-    num_outputs: usize,
-    gen: *mut (dyn Gen + Send),
-    start_node_at_sample: u64,
-}
-
-unsafe impl Send for Node {}
-
-impl Node {
-    pub fn new(_name: &'static str, gen: Box<dyn Gen + Send>) -> Self {
-        let num_outputs = gen.num_outputs();
-        let block_size = 0;
-        let output_buffers =
-            Box::<[Sample]>::into_raw(vec![0.0; num_outputs * block_size].into_boxed_slice());
-
-        let output_buffers_first_ptr = std::ptr::null_mut();
-
-        Node {
-            // name,
-            input_constants: Box::into_raw(
-                vec![0.0 as Sample; gen.num_inputs()].into_boxed_slice(),
-            ),
-            gen: Box::into_raw(gen),
-            output_buffers,
-            output_buffers_first_ptr,
-            num_outputs,
-            block_size,
-            start_node_at_sample: 0,
-        }
-    }
-    fn start_at_sample(&mut self, sample_time: u64) {
-        self.start_node_at_sample = sample_time;
-    }
-    // pub fn name(&self) -> &'static str {
-    //     self.name
-    // }
-    /// *Allocates memory*
-    /// Allocates enough memory for the given block size
-    pub fn init(&mut self, block_size: usize, sample_rate: Sample) {
-        // Free the previous buffer
-        unsafe {
-            drop(Box::from_raw(self.output_buffers));
-        }
-        self.output_buffers =
-            Box::<[Sample]>::into_raw(vec![0.0; self.num_outputs * block_size].into_boxed_slice());
-        self.output_buffers_first_ptr = if block_size * self.num_outputs > 0 {
-            // Get the pointer to the first f32 in the block without limiting its scope or going through a reference
-            self.output_buffers.cast::<f32>()
-        } else {
-            std::ptr::null_mut()
-        };
-        self.block_size = block_size;
-        unsafe {
-            (*self.gen).init(sample_rate);
-        }
-    }
-    /// Use the embedded Gen to generate values that are placed in the
-    /// output_buffer. The Graph will have already filled the input buffer with
-    /// the correct values.
-    #[inline]
-    pub fn process(
-        &mut self,
-        input_buffers: &NodeBufferRef,
-        sample_rate: Sample,
-        resources: &mut Resources,
-    ) -> GenState {
-        let mut outputs = NodeBufferRef {
-            buf: self.output_buffers_first_ptr,
-            num_channels: self.num_outputs,
-            block_size: self.block_size,
-            block_start_offset: 0,
-        };
-        let ctx = GenContext {
-            inputs: input_buffers,
-            outputs: &mut outputs,
-            sample_rate,
-        };
-        unsafe { (*self.gen).process(ctx, resources) }
-    }
-    pub fn set_constant(&mut self, value: Sample, input_index: usize) {
-        unsafe {
-            (*self.input_constants)[input_index] = value;
-        }
-    }
-    pub fn output_buffers(&self) -> NodeBufferRef {
-        NodeBufferRef {
-            buf: self.output_buffers_first_ptr,
-            num_channels: self.num_outputs,
-            block_size: self.block_size,
-            block_start_offset: 0,
-        }
-    }
-    pub fn num_inputs(&self) -> usize {
-        // self.gen.num_inputs()
-        // Not dynamic dispatch, may be faster
-        unsafe { &*self.input_constants }.len()
-    }
-    pub fn num_outputs(&self) -> usize {
-        self.num_outputs
-    }
-    pub fn input_indices_to_names(&self) -> Vec<&'static str> {
-        let mut list = vec![];
-        for i in 0..self.num_inputs() {
-            list.push(unsafe { (*self.gen).input_desc(i) });
-        }
-        list
-    }
-    pub fn output_indices_to_names(&self) -> Vec<&'static str> {
-        let mut list = vec![];
-        for i in 0..self.num_outputs() {
-            list.push(unsafe { (*self.gen).output_desc(i) });
-        }
-        list
-    }
-}
-impl Drop for Node {
-    fn drop(&mut self) {
-        drop(unsafe { Box::from_raw(self.gen) });
-        drop(unsafe { Box::from_raw(self.input_constants) });
-        drop(unsafe { Box::from_raw(self.output_buffers) });
-    }
-}
-
-#[derive(thiserror::Error, Debug, PartialEq)]
-pub enum RunGraphError {
-    #[error("Unable to create a node from the Graph: {0}")]
-    CouldNodeCreateNode(String),
-}
-
-pub struct RunGraphSettings {
-    /// How much time is added to every *relative* scheduling event to ensure the Change has time to travel to the GraphGen.
-    pub scheduling_latency: Duration,
-}
-impl Default for RunGraphSettings {
-    fn default() -> Self {
-        Self {
-            scheduling_latency: Duration::from_millis(50),
-        }
-    }
-}
-
-/// Wrapper around a [`Graph`] `Node` with convenience methods to run the
-/// Graph, either from an audio thread or for non-real time purposes.
-pub struct RunGraph {
-    graph_node: Node,
-    graph_sample_rate: Sample,
-    resources: Resources,
-    input_buffer_ptr: *mut f32,
-    input_buffer_length: usize,
-    input_node_buffer_ref: NodeBufferRef,
-    output_node_buffer_ref: NodeBufferRef,
-    resources_command_receiver: rtrb::Consumer<crate::ResourcesCommand>,
-    resources_response_sender: rtrb::Producer<crate::ResourcesResponse>,
-    /// The MusicalTimeMap that is shared between all sub Graphs running through
-    /// this RunGraph. Stored here so that the user can get access to changing it.
-    musical_time_map: Arc<RwLock<MusicalTimeMap>>,
-}
-
-impl RunGraph {
-    /// Prepare the necessary resources for running the graph. This will fail if
-    /// the Graph passed in has already been turned into a Node somewhere else,
-    /// for example if it has been pushed to a different Graph.
-    ///
-    /// Returns Self as well as ring buffer channels to apply changes to
-    /// Resources.
-    pub fn new(
-        graph: &mut Graph,
-        resources: Resources,
-        settings: RunGraphSettings,
-    ) -> Result<
-        (
-            Self,
-            rtrb::Producer<crate::ResourcesCommand>,
-            rtrb::Consumer<crate::ResourcesResponse>,
-        ),
-        RunGraphError,
-    > {
-        match graph.split_and_create_node() {
-            Ok(graph_node) => {
-                let input_buffer_length = graph_node.num_inputs() * graph_node.block_size;
-                let input_buffer_ptr = if input_buffer_length != 0 {
-                    let input_buffer = vec![0.0 as Sample; input_buffer_length].into_boxed_slice();
-                    let input_buffer = Box::into_raw(input_buffer);
-                    // Safety: we just created the slice of non-zero length
-                    input_buffer.cast::<f32>()
-                } else {
-                    std::ptr::null_mut()
-                };
-                let input_node_buffer_ref = NodeBufferRef {
-                    buf: input_buffer_ptr,
-                    num_channels: graph_node.num_inputs(),
-                    block_size: graph_node.block_size,
-                    block_start_offset: 0,
-                };
-                let graph_sample_rate = graph.sample_rate;
-                let musical_time_map = Arc::new(RwLock::new(MusicalTimeMap::new()));
-                //                TODO: Start the scheduler of the Graph
-                // Safety: The Nodes get initiated and their buffers allocated
-                // when the Graph is split and the Node is created from it.
-                // Therefore, we can safely store a reference to a buffer in
-                // that Node here. We want to store it to be able to return a
-                // reference to it instead of an owned value which includes a
-                // raw pointer.
-                let output_node_buffer_ref = graph_node.output_buffers();
-
-                let scheduler_start_time_stamp = Instant::now();
-                graph.start_scheduler(
-                    settings.scheduling_latency,
-                    scheduler_start_time_stamp,
-                    musical_time_map.clone(),
-                );
-                // Run a first update to make sure any queued changes get sent to the GraphGen
-                graph.update();
-                // Create ring buffer channels for communicating with Resources
-                let (resources_command_sender, resources_command_receiver) = RingBuffer::new(50);
-                let (resources_response_sender, resources_response_receiver) = RingBuffer::new(50);
-                Ok((
-                    Self {
-                        graph_node,
-                        graph_sample_rate,
-                        resources,
-                        input_buffer_length,
-                        input_buffer_ptr,
-                        input_node_buffer_ref,
-                        output_node_buffer_ref,
-                        musical_time_map,
-                        resources_command_receiver,
-                        resources_response_sender,
-                    },
-                    resources_command_sender,
-                    resources_response_receiver,
-                ))
-            }
-            Err(e) => Err(RunGraphError::CouldNodeCreateNode(e)),
-        }
-    }
-    /// Returns the input buffer that will be read as the input of the main
-    /// graph. For example, you may want to fill this with the sound inputs of
-    /// the sound card. The buffer does not get emptied automatically. If you
-    /// don't change it between calls to [`RunGraph::process_block`], its content will be static.
-    #[inline]
-    pub fn graph_input_buffers(&mut self) -> &mut NodeBufferRef {
-        &mut self.input_node_buffer_ref
-    }
-    pub fn run_resources_communication(&mut self, max_commands_to_process: usize) {
-        let mut i = 0;
-        while let Ok(command) = self.resources_command_receiver.pop() {
-            if let Err(e) = self
-                .resources_response_sender
-                .push(self.resources.apply_command(command))
-            {
-                eprintln!("Warning: A ResourcesResponse could not be sent back from the RunGraph. This may lead to dropping on the audio thread. {e}");
-            }
-            i += 1;
-            if i >= max_commands_to_process {
-                break;
-            }
-        }
-    }
-    /// Run the Graph for one block using the inputs currently stored in the
-    /// input buffer. The results can be accessed through the output buffer
-    /// through [`RunGraph::graph_output_buffers`].
-    pub fn process_block(&mut self) {
-        self.graph_node.process(
-            &self.input_node_buffer_ref,
-            self.graph_sample_rate,
-            &mut self.resources,
-        );
-    }
-    /// Return a reference to the buffer holding the output of the [`Graph`].
-    /// Channels which have no [`Connection`]/graph edge to them will be 0.0.
-    pub fn graph_output_buffers(&self) -> &NodeBufferRef {
-        &self.output_node_buffer_ref
-    }
-    pub fn block_size(&self) -> usize {
-        self.output_node_buffer_ref.block_size()
-    }
-}
-
-impl Drop for RunGraph {
-    fn drop(&mut self) {
-        // Safety: The slice is allocated when the RunGraph is created with the
-        // given size, unless that size is 0 in which case no allocation is
-        // made.
-        if !self.input_buffer_ptr.is_null() {
-            unsafe {
-                drop(Box::from_raw(std::slice::from_raw_parts_mut(
-                    self.input_buffer_ptr,
-                    self.input_buffer_length,
-                )));
-            }
-        }
-    }
-}
-unsafe impl Send for RunGraph {}
 
 /// Buffers the output of a node from last block to simplify feedback nodes and
 /// make sure they work in all possible graphs.
@@ -4361,897 +3247,5 @@ impl Gen for NaiveSine {
     }
 }
 
-use rtrb::RingBuffer;
-use slotmap::{new_key_type, SecondaryMap, SlotMap};
 #[cfg(test)]
-mod tests {
-
-    use std::sync::atomic::AtomicBool;
-
-    use crate::{
-        buffer::{Buffer, BufferReader},
-        controller::Controller,
-        ResourcesSettings,
-    };
-
-    use super::*;
-    // Outputs its input value + 1
-    struct OneGen {}
-    impl Gen for OneGen {
-        fn process(&mut self, ctx: GenContext, _resources: &mut Resources) -> GenState {
-            for i in 0..ctx.block_size() {
-                ctx.outputs.write(ctx.inputs.read(0, i) + 1.0, 0, i);
-            }
-            GenState::Continue
-        }
-        fn input_desc(&self, input: usize) -> &'static str {
-            match input {
-                0 => "passthrough",
-                _ => "",
-            }
-        }
-
-        fn num_inputs(&self) -> usize {
-            1
-        }
-
-        fn num_outputs(&self) -> usize {
-            1
-        }
-    }
-    struct DummyGen {
-        counter: f32,
-    }
-    impl Gen for DummyGen {
-        fn process(&mut self, ctx: GenContext, _resources: &mut Resources) -> GenState {
-            for i in 0..ctx.block_size() {
-                self.counter += 1.;
-                ctx.outputs
-                    .write(ctx.inputs.read(0, i) + self.counter, 0, i);
-            }
-            GenState::Continue
-        }
-
-        fn input_desc(&self, input: usize) -> &'static str {
-            match input {
-                0 => "counter",
-                _ => "",
-            }
-        }
-
-        fn num_inputs(&self) -> usize {
-            1
-        }
-
-        fn num_outputs(&self) -> usize {
-            1
-        }
-    }
-    fn test_resources_settings() -> ResourcesSettings {
-        ResourcesSettings::default()
-    }
-    fn test_run_graph(graph: &mut Graph, settings: RunGraphSettings) -> RunGraph {
-        let resources = Resources::new(test_resources_settings());
-        let (run_graph, _, _) = RunGraph::new(graph, resources, settings).unwrap();
-        run_graph
-    }
-
-    #[test]
-    fn create_graph() {
-        let mut graph: Graph = Graph::new(GraphSettings::default());
-        let node_id = graph.push(DummyGen { counter: 0.0 });
-        graph.connect(Connection::graph_output(&node_id)).unwrap();
-        graph.init();
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 31), 32.0);
-    }
-    #[test]
-    fn multiple_nodes() {
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: 4,
-            ..Default::default()
-        });
-        let node1 = DummyGen { counter: 0.0 };
-        let node2 = DummyGen { counter: 0.0 };
-        let node3 = DummyGen { counter: 0.0 };
-        let node_id1 = graph.push(node1);
-        let node_id2 = graph.push(node2);
-        let node_id3 = graph.push(node3);
-        graph.connect(node_id1.to(&node_id3)).unwrap();
-        graph.connect(node_id2.to(&node_id3)).unwrap();
-        graph.connect(Connection::graph_output(&node_id3)).unwrap();
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 2), 9.0);
-    }
-    #[test]
-    fn constant_inputs() {
-        const BLOCK: usize = 16;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let node = DummyGen { counter: 0.0 };
-        let node_id = graph.push(node);
-        graph.connect(constant(0.5).to(&node_id)).unwrap();
-        graph.connect(Connection::graph_output(&node_id)).unwrap();
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, BLOCK - 1), 16.5);
-    }
-    #[test]
-    fn graph_in_a_graph() {
-        const BLOCK: usize = 16;
-        const CONSTANT_INPUT_TO_NODE: Sample = 0.25;
-        let mut inner_graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let node = DummyGen { counter: 0.0 };
-        let node_address = inner_graph.push(node);
-        inner_graph
-            .connect(Connection::graph_output(&node_address))
-            .unwrap();
-        let mut top_level_graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let inner_graph_node_id = top_level_graph.push(inner_graph);
-        top_level_graph
-            .connect(Connection::graph_output(&inner_graph_node_id).channels(2))
-            .unwrap();
-        // Parent graphs should propagate a connection to its child graphs
-        // without issue, but it will be scheduled instead of applied directly
-        top_level_graph
-            .connect(constant(CONSTANT_INPUT_TO_NODE).to(&node_address))
-            .unwrap();
-        let mut run_graph = test_run_graph(&mut top_level_graph, RunGraphSettings::default());
-        run_graph.process_block();
-        assert_eq!(
-            run_graph.graph_output_buffers().read(0, BLOCK - 2),
-            (BLOCK - 1) as Sample + CONSTANT_INPUT_TO_NODE
-        );
-    }
-    #[test]
-    fn feedback_in_graph() {
-        // v-------------<
-        // |   2 -> 3 -> 4
-        // |   |    |
-        // fbv fbv fbv
-        //   0     -> 1 -> out
-        const BLOCK: usize = 4;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let n0 = graph.push(DummyGen { counter: 0.0 });
-        let n1 = graph.push(DummyGen { counter: 0.0 });
-        let n2 = graph.push(DummyGen { counter: 4.0 });
-        let n3 = graph.push(DummyGen { counter: 5.0 });
-        let n4 = graph.push(DummyGen { counter: 1.0 });
-        graph.connect(Connection::graph_output(&n1)).unwrap();
-        graph
-            .connect(Connection::graph_output(&n0).to_index(1))
-            .unwrap();
-        graph.connect(n0.to(&n1)).unwrap();
-        graph.connect(n2.to(&n3)).unwrap();
-        graph.connect(n3.to(&n4)).unwrap();
-        graph.connect(n2.to(&n0).feedback(true)).unwrap();
-        graph.connect(n3.to(&n1).feedback(true)).unwrap();
-        graph.connect(n4.to(&n0).feedback(true)).unwrap();
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        run_graph.process_block();
-        let block1 = vec![2.0f32, 4., 6., 8.];
-        let block2 = vec![39f32, 47., 55., 63.];
-        for (&output, expected) in run_graph
-            .graph_output_buffers()
-            .get_channel(0)
-            .iter()
-            .zip(block1)
-        {
-            assert_eq!(
-                output,
-                expected,
-                "block1 failed with o: {}, expected: {}, n1: {:#?}, n0: {:#?}",
-                output,
-                expected,
-                run_graph.graph_output_buffers().get_channel(0),
-                run_graph.graph_output_buffers().get_channel(1)
-            );
-        }
-        run_graph.process_block();
-        for (&output, expected) in run_graph
-            .graph_output_buffers()
-            .get_channel(0)
-            .iter()
-            .zip(block2)
-        {
-            assert_eq!(
-                output,
-                expected,
-                "block2 failed with o: {}, expected: {}, n1: {:#?}, n0: {:#?}",
-                output,
-                expected,
-                run_graph.graph_output_buffers().get_channel(0),
-                run_graph.graph_output_buffers().get_channel(1)
-            );
-        }
-    }
-
-    #[test]
-    fn changing_the_graph() {
-        const BLOCK: usize = 1;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let triangular_sequence = |n| (n * (n + 1)) / 2;
-        for i in 0..10 {
-            let node = DummyGen { counter: 0.0 };
-            let node_id = graph.push(node);
-            graph.connect(constant(0.5).to(&node_id)).unwrap();
-            graph.connect(Connection::graph_output(&node_id)).unwrap();
-            graph.update();
-            run_graph.process_block();
-            assert_eq!(
-                run_graph.graph_output_buffers().read(0, 0),
-                (i + 1) as f32 * 0.5 + triangular_sequence(i + 1) as f32,
-                "i: {}, output: {}, expected: {}",
-                i,
-                run_graph.graph_output_buffers().read(0, 0),
-                (i + 1) as f32 * 0.5 + triangular_sequence(i + 1) as f32,
-            );
-        }
-    }
-
-    #[test]
-    fn graph_inputs() {
-        const BLOCK: usize = 1;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            num_inputs: 3,
-            ..Default::default()
-        });
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let node = DummyGen { counter: 0.0 };
-        let node_id = graph.push(node);
-        graph
-            .connect(
-                Connection::graph_input(&node_id)
-                    .from_index(2)
-                    .to_label("counter"),
-            )
-            .unwrap();
-        graph.connect(Connection::graph_output(&node_id)).unwrap();
-        graph.update();
-        run_graph.graph_input_buffers().fill_channel(10.0, 2);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 11.0);
-    }
-
-    #[test]
-    fn remove_nodes() {
-        const BLOCK: usize = 1;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let mut nodes = vec![];
-        let mut last_node = None;
-        for _ in 0..10 {
-            let node = graph.push(OneGen {});
-            if let Some(last) = last_node.take() {
-                graph.connect(node.to(&last)).unwrap();
-            } else {
-                graph.connect(Connection::graph_output(&node)).unwrap();
-            }
-            last_node = Some(node.clone());
-            nodes.push(node);
-        }
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 10.0);
-        // Convert NodeAddress to RawNodeAddress for removal, which we know will work because we pushed it straight to the graph
-        graph.free_node(nodes[9].to_raw().unwrap()).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 9.0);
-        graph.free_node(nodes[4].to_raw().unwrap()).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 4.0);
-        graph.connect(nodes[5].to(&nodes[3])).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 8.0);
-        for node in &nodes[1..4] {
-            graph.free_node(node.to_raw().unwrap()).unwrap();
-        }
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 1.0);
-        assert_eq!(
-            graph.free_node(nodes[4].to_raw().unwrap()),
-            Err(FreeError::NodeNotFound)
-        );
-    }
-
-    #[test]
-    fn parallel_mutation() {
-        // Here we just want to check that the program doesn't crash/segfault when changing the graph while running the GraphGen.
-        const BLOCK: usize = 1;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let mut nodes = vec![];
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let mut last_node = None;
-        let done_flag = Arc::new(AtomicBool::new(false));
-        let audio_done_flag = done_flag.clone();
-        let audio_thread = std::thread::spawn(move || {
-            while !audio_done_flag.load(Ordering::SeqCst) {
-                run_graph.process_block();
-                std::thread::sleep(std::time::Duration::from_millis(1));
-            }
-        });
-        for _ in 0..10 {
-            let node = graph.push(OneGen {});
-            if let Some(last) = last_node.take() {
-                graph.connect(node.to(&last)).unwrap();
-            } else {
-                graph.connect(Connection::graph_output(&node)).unwrap();
-            }
-            last_node = Some(node.clone());
-            nodes.push(node);
-            graph.update();
-            std::thread::sleep(std::time::Duration::from_millis(3));
-        }
-        for node in nodes.into_iter().rev() {
-            graph.free_node(node.to_raw().unwrap()).unwrap();
-            graph.commit_changes();
-            std::thread::sleep(std::time::Duration::from_millis(2));
-        }
-        let mut nodes = vec![];
-        last_node = None;
-        for _ in 0..10 {
-            let node = graph.push(DummyGen { counter: 0. });
-            if let Some(last) = last_node.take() {
-                graph.connect(node.to(&last)).unwrap();
-            } else {
-                graph.connect(Connection::graph_output(&node)).unwrap();
-            }
-            last_node = Some(node.clone());
-            nodes.push(node);
-            graph.update();
-            std::thread::sleep(std::time::Duration::from_millis(1));
-        }
-        use rand::seq::SliceRandom;
-        let mut rng = rand::thread_rng();
-        nodes.shuffle(&mut rng);
-        for node in nodes.into_iter() {
-            graph.free_node(node.to_raw().unwrap()).unwrap();
-            graph.update();
-            std::thread::sleep(std::time::Duration::from_millis(1));
-        }
-        drop(graph);
-        done_flag.store(true, Ordering::SeqCst);
-        audio_thread.join().unwrap();
-    }
-    struct SelfFreeing {
-        samples_countdown: usize,
-        value: Sample,
-        mend: bool,
-    }
-    impl Gen for SelfFreeing {
-        fn process(&mut self, ctx: GenContext, _resources: &mut Resources) -> GenState {
-            for i in 0..ctx.block_size() {
-                if self.samples_countdown == 0 {
-                    if self.mend {
-                        ctx.outputs.write(ctx.inputs.read(0, i), 0, i);
-                    } else {
-                        ctx.outputs.write(0., 0, i);
-                    }
-                } else {
-                    ctx.outputs.write(ctx.inputs.read(0, i) + self.value, 0, i);
-                    self.samples_countdown -= 1;
-                }
-            }
-            if self.samples_countdown == 0 {
-                if self.mend {
-                    GenState::FreeSelfMendConnections
-                } else {
-                    GenState::FreeSelf
-                }
-            } else {
-                GenState::Continue
-            }
-        }
-
-        fn num_inputs(&self) -> usize {
-            1
-        }
-
-        fn num_outputs(&self) -> usize {
-            1
-        }
-    }
-
-    #[test]
-    fn self_freeing_nodes() {
-        const BLOCK: usize = 1;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            ..Default::default()
-        });
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let n0 = graph.push(SelfFreeing {
-            samples_countdown: 2,
-            value: 1.,
-            mend: true,
-        });
-        let n1 = graph.push(SelfFreeing {
-            samples_countdown: 1,
-            value: 1.,
-            mend: true,
-        });
-        let n2 = graph.push(SelfFreeing {
-            samples_countdown: 4,
-            value: 2.,
-            mend: false,
-        });
-        let n3 = graph.push(SelfFreeing {
-            samples_countdown: 5,
-            value: 3.,
-            mend: false,
-        });
-        graph.connect(Connection::graph_output(&n0)).unwrap();
-        graph.connect(n1.to(&n0)).unwrap();
-        graph.connect(n2.to(&n1)).unwrap();
-        graph.connect(n3.to(&n2)).unwrap();
-
-        graph.update();
-        assert_eq!(graph.num_nodes(), 4);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 7.0);
-        graph.update();
-        // Still 4 since the node has been added to the free node queue, but there
-        // hasn't been a new generation in the GraphGen yet.
-        assert_eq!(graph.num_nodes(), 4);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 6.0);
-        graph.update();
-        assert_eq!(graph.num_nodes(), 3);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 5.0);
-        graph.update();
-        assert_eq!(graph.num_nodes(), 2);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 5.0);
-        graph.update();
-        assert_eq!(graph.num_nodes(), 2);
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 0.0);
-        graph.update();
-        assert_eq!(graph.num_nodes(), 1);
-        run_graph.process_block();
-        graph.update();
-        assert_eq!(graph.num_nodes(), 0);
-    }
-    #[test]
-    fn scheduling() {
-        const BLOCK: usize = 4;
-        const SR: u64 = 44100;
-        let mut graph: Graph = Graph::new(GraphSettings {
-            block_size: BLOCK,
-            sample_rate: SR as f32,
-            ..Default::default()
-        });
-        // let mut graph_node = graph_node(&mut graph);
-        let mut run_graph = test_run_graph(
-            &mut graph,
-            RunGraphSettings {
-                scheduling_latency: Duration::from_millis(0),
-            },
-        );
-        let node0 = graph.push(OneGen {});
-        let node1 = graph.push(OneGen {});
-        graph.connect(Connection::graph_output(&node0)).unwrap();
-        graph.connect(node1.to(&node0)).unwrap();
-        // This gets applied only one sample late. Why?
-        graph.connect(constant(2.).to(&node1)).unwrap();
-        graph.update();
-        graph
-            .schedule_change(
-                ParameterChange::superseconds(
-                    node0.clone(),
-                    1.0,
-                    Superseconds::from_samples(3, SR),
-                )
-                .i(0),
-            )
-            .unwrap();
-        graph
-            .schedule_change(
-                ParameterChange::superseconds(
-                    node0.clone(),
-                    2.0,
-                    Superseconds::from_samples(2, SR),
-                )
-                .l("passthrough"),
-            )
-            .unwrap();
-        graph
-            .schedule_change(
-                ParameterChange::superseconds(
-                    node1.clone(),
-                    10.0,
-                    Superseconds::from_samples(7, SR),
-                )
-                .i(0),
-            )
-            .unwrap();
-        // Schedule far into the future, this should not show up in the test output
-        graph
-            .schedule_change(
-                ParameterChange::duration_from_now(node1.clone(), 3000.0, Duration::from_secs(100))
-                    .i(0),
-            )
-            .unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 4.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 1), 4.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 2), 6.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 3), 5.0);
-        graph
-            .schedule_change(
-                ParameterChange::superseconds(node0, 0.0, Superseconds::from_samples(5, SR)).i(0),
-            )
-            .unwrap();
-        graph
-            .schedule_change(
-                ParameterChange::superseconds(
-                    node1.clone(),
-                    0.0,
-                    Superseconds::from_samples(6, SR),
-                )
-                .l("passthrough"),
-            )
-            .unwrap();
-        assert_eq!(
-            graph.schedule_change(
-                ParameterChange::superseconds(
-                    node1.clone(),
-                    100.0,
-                    Superseconds::from_samples(6, SR)
-                )
-                .l("pasta")
-            ),
-            Err(ScheduleError::InputLabelNotFound("pasta"))
-        );
-        graph.update();
-        run_graph.process_block();
-        dbg!(run_graph.graph_output_buffers().get_channel(0));
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 5.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 1), 4.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 2), 2.0);
-        assert_eq!(run_graph.graph_output_buffers().read(0, 3), 12.0);
-        // Schedule "now", but this will be a few samples into the
-        // future depending on the time it takes to run this test.
-        graph
-            .schedule_change(ParameterChange::now(node1, 1000.0).i(0))
-            .unwrap();
-        graph.update();
-        // Move a few hundred samples into the future
-        for _ in 0..600 {
-            run_graph.process_block();
-            graph.update();
-        }
-        // If this fails, it may be because the machine is too slow. Try
-        // increasing the number of iterations above.
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 1002.0);
-    }
-    #[test]
-    fn index_routing() {
-        let graph_settings = GraphSettings {
-            block_size: 4,
-            sample_rate: 44100.,
-            num_outputs: 2,
-            ..Default::default()
-        };
-        let mut graph = Graph::new(graph_settings);
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let mult = graph.push(Mult);
-        let five = graph.push(
-            gen(move |ctx, _resources| {
-                let out_chan = ctx.outputs.split_mut().next().unwrap();
-                for o0 in out_chan {
-                    *o0 = 5.;
-                }
-                GenState::Continue
-            })
-            .output("o"),
-        );
-        let nine = graph.push(
-            gen(move |ctx, _resources| {
-                let out_chan = ctx.outputs.split_mut().next().unwrap();
-                for o0 in out_chan {
-                    *o0 = 9.;
-                }
-                GenState::Continue
-            })
-            .output("o"),
-        );
-        // Connecting the node to the graph output
-        graph.connect(mult.to_graph_out()).unwrap();
-        // Multiply 5 by 9
-        graph.connect(five.to(&mult).to_index(0)).unwrap();
-        graph.connect(nine.to(&mult).to_index(1)).unwrap();
-        // You need to commit changes and update if the graph is running.
-        graph.commit_changes();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 45.0);
-    }
-    #[test]
-    fn index_routing_advanced() {
-        let graph_settings = GraphSettings {
-            block_size: 4,
-            sample_rate: 44100.,
-            num_outputs: 2,
-            ..Default::default()
-        };
-        let mut graph = Graph::new(graph_settings);
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        let multiplier = graph.push(
-            gen(move |ctx, _resources| {
-                for i in 0..ctx.block_size() {
-                    let i0 = ctx.inputs.read(0, i) * 1.;
-                    let i1 = ctx.inputs.read(1, i) * 2.;
-                    let i2 = ctx.inputs.read(2, i) * 3.;
-                    let i3 = ctx.inputs.read(3, i) * 4.;
-                    let i4 = ctx.inputs.read(4, i) * 5.;
-                    let i5 = ctx.inputs.read(5, i) * 6.;
-                    let mut value = i0;
-                    for i in [i1, i2, i3, i4, i5] {
-                        if i != 0.0 {
-                            value *= i;
-                        }
-                    }
-                    ctx.outputs.write(value, 0, i);
-                }
-                GenState::Continue
-            })
-            .output("o")
-            .input("i0")
-            .input("i1")
-            .input("i2")
-            .input("i3")
-            .input("i4")
-            .input("i5"),
-        );
-        let five = graph.push(
-            gen(move |ctx, _resources| {
-                for o0 in ctx.outputs.split_mut().next().unwrap() {
-                    *o0 = 5.;
-                }
-                GenState::Continue
-            })
-            .output("o"),
-        );
-        let nine = graph.push(
-            gen(move |ctx, _resources| {
-                for o0 in ctx.outputs.split_mut().next().unwrap() {
-                    *o0 = 9.;
-                }
-                GenState::Continue
-            })
-            .output("o"),
-        );
-        let numberer = graph.push(
-            gen(move |ctx, _resources| {
-                for i in 0..ctx.block_size() {
-                    ctx.outputs.write(0.0, 0, i);
-                    ctx.outputs.write(1.0, 1, i);
-                    ctx.outputs.write(2.0, 2, i);
-                    ctx.outputs.write(3.0, 3, i);
-                    ctx.outputs.write(4.0, 4, i);
-                    ctx.outputs.write(5.0, 5, i);
-                    ctx.outputs.write(6.0, 6, i);
-                    ctx.outputs.write(7.0, 7, i);
-                    ctx.outputs.write(8.0, 8, i);
-                    ctx.outputs.write(9.0, 9, i);
-                }
-                GenState::Continue
-            })
-            .output("o0")
-            .output("o1")
-            .output("o2")
-            .output("o3")
-            .output("o4")
-            .output("o5")
-            .output("o6")
-            .output("o7")
-            .output("o8")
-            .output("o9"),
-        );
-        // Connecting the node to the graph output
-        graph.connect(multiplier.to_graph_out()).unwrap();
-        // Multiply 5 by 9
-        graph.connect(nine.to(&multiplier).to_index(0)).unwrap();
-        graph.connect(five.to(&multiplier).to_index(1)).unwrap();
-        // You need to commit changes and update if the graph is running.
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 90.0);
-
-        graph.disconnect(five.to(&multiplier).to_index(1)).unwrap();
-        graph.connect(five.to(&multiplier).to_index(3)).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 180.0);
-
-        graph.connect(five.to(&multiplier).to_index(4)).unwrap();
-        graph.connect(five.to(&multiplier).to_label("i5")).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 135000.0);
-
-        graph.connect(Connection::clear_to_nodes(&five)).unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 9.0);
-
-        graph
-            .connect(numberer.to(&multiplier).from_index(2).to_index(4))
-            .unwrap();
-        graph
-            .connect(numberer.to(&multiplier).from_index(7).to_index(2))
-            .unwrap();
-        graph.update();
-        run_graph.process_block();
-        assert_eq!(run_graph.graph_output_buffers().read(0, 0), 1890.0);
-    }
-    #[test]
-    fn sending_buffer_to_resources() {
-        const BLOCK_SIZE: usize = 16;
-        let graph_settings = GraphSettings {
-            block_size: BLOCK_SIZE,
-            sample_rate: 44100.,
-            num_outputs: 2,
-            ..Default::default()
-        };
-        let mut graph = Graph::new(graph_settings);
-        let resources = Resources::new(ResourcesSettings {
-            max_wavetables: 0,
-            max_buffers: 3,
-            max_user_data: 0,
-        });
-        let (mut run_graph, resources_command_sender, resources_response_receiver) = RunGraph::new(
-            &mut graph,
-            resources,
-            RunGraphSettings {
-                scheduling_latency: Duration::from_secs(0),
-            },
-        )
-        .unwrap();
-
-        let mut controller = Controller::new(
-            graph,
-            |e| println!("{e}"),
-            resources_command_sender,
-            resources_response_receiver,
-        );
-        let mut k = controller.get_knyst_commands();
-
-        // Send a buffer to the RunGraph
-        let buffer_vec = vec![10., 9., 8., 7., 6., 5., 4., 3., 2., 1., 0.];
-        let buffer = Buffer::from_vec(buffer_vec.clone(), 44100.);
-        let buffer_id = k.insert_buffer(buffer);
-        // Push a buffer reader that can read the Buffer
-        let mut buffer_reader = BufferReader::new(
-            crate::IdOrKey::Id(buffer_id),
-            1.0,
-            crate::StopAction::FreeSelf,
-        );
-        buffer_reader.looping = true;
-        let br = k.push(buffer_reader);
-        k.connect(br.to_graph_out());
-        // Process the Controller and RunGraph in order
-        controller.run(300);
-        run_graph.run_resources_communication(50);
-        run_graph.process_block();
-        // We should be able to read the buffer
-        let mut cycling_vec = buffer_vec.iter().cycle();
-        for i in 0..BLOCK_SIZE {
-            assert_eq!(
-                run_graph.graph_output_buffers().read(0, i),
-                *cycling_vec.next().unwrap()
-            );
-        }
-    }
-    #[test]
-    fn start_nodes_with_sample_precision() {
-        const SR: u64 = 44100;
-        const BLOCK_SIZE: usize = 8;
-        let graph_settings = GraphSettings {
-            block_size: BLOCK_SIZE,
-            sample_rate: SR as f32,
-            num_outputs: 2,
-            ..Default::default()
-        };
-        let mut graph = Graph::new(graph_settings);
-        let mut counter0 = 0.;
-        // push before starting the graph
-        let n0 = graph.push_at_time(
-            gen(move |ctx, _resources| {
-                let out_chan = ctx.outputs.split_mut().next().unwrap();
-                for sample in out_chan {
-                    *sample = counter0;
-                    counter0 += 1.0;
-                }
-                GenState::Continue
-            })
-            .output("out"),
-            Superseconds::from_samples(1, SR),
-        );
-        graph.connect(n0.to_graph_out()).unwrap();
-        let mut counter1 = 0.;
-        let n1 = graph.push_at_time(
-            gen(move |ctx, _resources| {
-                for i in 0..ctx.outputs.block_size() {
-                    ctx.outputs.write(counter1 * -1. - 1., 0, i);
-                    counter1 += 1.0;
-                }
-                GenState::Continue
-            })
-            .output("out"),
-            Superseconds::from_samples(2, SR),
-        );
-        graph.connect(n1.to_graph_out()).unwrap();
-        let mut run_graph = test_run_graph(&mut graph, RunGraphSettings::default());
-        run_graph.process_block();
-
-        for i in 0..BLOCK_SIZE {
-            assert_eq!(run_graph.graph_output_buffers().get_channel(0)[i], 0.)
-        }
-        // push after starting the graph
-        let mut counter2 = 0.;
-        let n2 = graph.push_at_time(
-            gen(move |ctx, _resources| {
-                for i in 0..ctx.outputs.block_size() {
-                    ctx.outputs.write(counter2, 0, i);
-                    counter2 += 1.0;
-                }
-                GenState::Continue
-            })
-            .output("out"),
-            Superseconds::from_samples(3 + BLOCK_SIZE as u64, SR),
-        );
-        let mut counter3 = 0.;
-        let n3 = graph.push_at_time(
-            gen(move |ctx, _resources| {
-                for i in 0..ctx.outputs.block_size() {
-                    ctx.outputs.write(counter3 * -1. - 1., 0, i);
-                    counter3 += 1.0;
-                }
-                GenState::Continue
-            })
-            .output("out"),
-            Superseconds::from_samples(4 + BLOCK_SIZE as u64, SR),
-        );
-        graph.connect(n2.to_graph_out()).unwrap();
-        graph.connect(n3.to_graph_out()).unwrap();
-        graph.update();
-        for _ in 0..2 {
-            run_graph.process_block();
-            for i in 0..BLOCK_SIZE {
-                assert_eq!(run_graph.graph_output_buffers().get_channel(0)[i], 0.)
-            }
-        }
-    }
-}
+mod tests;
