+++ src/prelude.rs
@@ -1,8 +1,8 @@
 pub use crate::audio_backend::AudioBackend;
 pub use crate::buffer::{Buffer, BufferKey, BufferReader};
 pub use crate::graph::{
-    constant, gen, Connection, GenContext, GenState, Graph, GraphInput, GraphSettings, Mult,
-    PanMonoToStereo, ParameterChange, Ramp, RunGraphSettings,
+    connection::constant, gen, Connection, GenContext, GenState, Graph, GraphInput, GraphSettings,
+    Mult, PanMonoToStereo, ParameterChange, Ramp, RunGraphSettings,
 };
 pub use crate::wavetable::{Wavetable, WavetableKey, TABLE_POWER, TABLE_SIZE};
 pub use crate::{AnyData, Resources, ResourcesSettings, Sample, StopAction};
